+++ libs/langgraph/src/pregel/loop.ts
@@ -49,9 +49,7 @@ import {
 import {
   gatherIterator,
   gatherIteratorSync,
-  IterableReadableWritableStream,
   prefixGenerator,
-  StreamChunk,
 } from "../utils.js";
 import {
   mapCommand,
@@ -82,6 +80,9 @@ const INPUT_DONE = Symbol.for("INPUT_DONE");
 const INPUT_RESUMING = Symbol.for("INPUT_RESUMING");
 const DEFAULT_LOOP_LIMIT = 25;
 
+// [namespace, streamMode, payload]
+export type StreamChunk = [string[], StreamMode, unknown];
+
 export type PregelLoopInitializeParams = {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   input?: any | Command;
@@ -100,6 +101,68 @@ export type PregelLoopInitializeParams = {
   manager?: CallbackManagerForChainRun;
   debug: boolean;
 };
+
+export class IterableReadableWritableStream extends IterableReadableStream<StreamChunk> {
+  modes: Set<StreamMode>;
+
+  private controller: ReadableStreamDefaultController;
+
+  private passthroughFn?: (chunk: StreamChunk) => void;
+
+  private _closed: boolean = false;
+
+  get closed() {
+    return this._closed;
+  }
+
+  constructor(params: {
+    passthroughFn?: (chunk: StreamChunk) => void;
+    modes: Set<StreamMode>;
+  }) {
+    let streamControllerPromiseResolver: (
+      controller: ReadableStreamDefaultController
+    ) => void;
+    const streamControllerPromise: Promise<ReadableStreamDefaultController> =
+      new Promise<ReadableStreamDefaultController>((resolve) => {
+        streamControllerPromiseResolver = resolve;
+      });
+
+    super({
+      start: (controller) => {
+        streamControllerPromiseResolver!(controller);
+      },
+    });
+
+    // .start() will always be called before the stream can be interacted
+    // with anyway
+    void streamControllerPromise.then((controller) => {
+      this.controller = controller;
+    });
+
+    this.passthroughFn = params.passthroughFn;
+    this.modes = params.modes;
+  }
+
+  push(chunk: StreamChunk) {
+    this.passthroughFn?.(chunk);
+    this.controller.enqueue(chunk);
+  }
+
+  close() {
+    try {
+      this.controller.close();
+    } catch (e) {
+      // pass
+    } finally {
+      this._closed = true;
+    }
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  error(e: any) {
+    this.controller.error(e);
+  }
+}
 
 type PregelLoopParams = {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
