+++ libs/langgraph/src/utils.ts
@@ -6,9 +6,7 @@ import {
   RunnableConfig,
 } from "@langchain/core/runnables";
 import { AsyncLocalStorageProviderSingleton } from "@langchain/core/singletons";
-import { IterableReadableStream } from "@langchain/core/utils/stream";
 import { ensureLangGraphConfig } from "./pregel/utils/config.js";
-import { StreamMode } from "./pregel/types.js";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export interface RunnableCallableArgs extends Partial<any> {
@@ -180,156 +178,8 @@ export function patchConfigurable(
   }
 }
 
-// [namespace, streamMode, payload]
-export type StreamChunk = [string[], StreamMode, unknown];
-
-export class IterableReadableWritableStream extends IterableReadableStream<StreamChunk> {
-  modes: Set<StreamMode>;
-
-  private controller: ReadableStreamDefaultController;
-
-  private passthroughFn?: (chunk: StreamChunk) => void;
-
-  private _closed: boolean = false;
-
-  get closed() {
-    return this._closed;
-  }
-
-  constructor(params: {
-    passthroughFn?: (chunk: StreamChunk) => void;
-    modes: Set<StreamMode>;
-  }) {
-    let streamControllerPromiseResolver: (
-      controller: ReadableStreamDefaultController
-    ) => void;
-    const streamControllerPromise: Promise<ReadableStreamDefaultController> =
-      new Promise<ReadableStreamDefaultController>((resolve) => {
-        streamControllerPromiseResolver = resolve;
-      });
-
-    super({
-      start: (controller) => {
-        streamControllerPromiseResolver!(controller);
-      },
-    });
-
-    // .start() will always be called before the stream can be interacted
-    // with anyway
-    void streamControllerPromise.then((controller) => {
-      this.controller = controller;
-    });
-
-    this.passthroughFn = params.passthroughFn;
-    this.modes = params.modes;
-  }
-
-  push(chunk: StreamChunk) {
-    this.passthroughFn?.(chunk);
-    this.controller.enqueue(chunk);
-  }
-
-  close() {
-    try {
-      this.controller.close();
-    } catch (e) {
-      // pass
-    } finally {
-      this._closed = true;
-    }
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  error(e: any) {
-    this.controller.error(e);
-  }
-}
-
-export class StreamFlusher {
-  private _nextChunk: Promise<IteratorResult<StreamChunk>> | undefined;
-
-  private readonly _stream: IterableReadableWritableStream;
-
-  private readonly _streamMode: StreamMode[];
-
-  private readonly _streamSubgraphs?: boolean;
-
-  private readonly _streamModeSingle?: boolean;
-
-  terminated: boolean = false;
-
-  private readonly _timeoutSymbol: symbol = Symbol.for("timeout");
-
-  constructor({
-    stream,
-    streamMode,
-    streamModeSingle,
-    streamSubgraphs,
-  }: {
-    stream: IterableReadableWritableStream;
-    streamMode: StreamMode[];
-    streamSubgraphs?: boolean;
-    streamModeSingle?: boolean;
-  }) {
-    this._stream = stream;
-    this._streamMode = streamMode;
-    this._streamModeSingle = streamModeSingle;
-    this._streamSubgraphs = streamSubgraphs;
-  }
-
-  async *flushAll() {
-    yield* this.flush({ blockUntilFinished: true });
-  }
-
-  async *flush(options?: { blockUntilFinished: boolean }) {
-    const { blockUntilFinished = false } = options ?? {};
-
-    if (this.terminated) {
-      return;
-    }
-
-    while (!this.terminated) {
-      if (!this._nextChunk) {
-        this._nextChunk = this._stream.next();
-      }
-
-      const chunk = blockUntilFinished
-        ? await this._nextChunk
-        : await Promise.race([this._nextChunk, this._timeout()]);
-      if (chunk === this._timeoutSymbol) {
-        break;
-      } else {
-        this._nextChunk = undefined;
-      }
-
-      const chunkResult = chunk as IteratorResult<StreamChunk>;
-
-      if (chunkResult.done) {
-        this.terminated = true;
-        return chunkResult.value;
-      } else {
-        if (chunkResult.value === undefined) {
-          throw new Error("Data structure error.");
-        }
-        const [namespace, mode, payload] = chunkResult.value;
-        if (this._streamMode.includes(mode)) {
-          if (this._streamSubgraphs && !this._streamModeSingle) {
-            yield [namespace, mode, payload];
-          } else if (!this._streamModeSingle) {
-            yield [mode, payload];
-          } else if (this._streamSubgraphs) {
-            yield [namespace, payload];
-          } else {
-            yield payload;
-          }
-        }
-      }
-    }
-  }
-
-  private _timeout() {
-    return new Promise((resolve) => {
-      process.nextTick(() => resolve(this._timeoutSymbol));
-    }) as Promise<typeof this._timeoutSymbol>;
-  }
-}
+// TODO: name this something sensible, and move to utils
+export type Promisified<T extends (...args: unknown[]) => unknown> =
+  ReturnType<T> extends Promise<unknown>
+    ? ReturnType<T>
+    : Promise<ReturnType<T>>;
